WIP: Add differential drive node
